FT_ATOI -> LONG FOR CHECK INT

+++ includes/Libft/src/ft_to/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: abesneux <abesneux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 15:25:09 by abesneux          #+#    #+#             */
-/*   Updated: 2024/01/12 18:44:59 by abesneux         ###   ########.fr       */
+/*   Updated: 2024/03/21 12:19:46 by abesneux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	sign;
-	int	result;
+	long	result;
 
 	sign = 1;
 	result = 0;
